backend/requirement.txt
langchain
fastmcp
python-dotenv